backend/requirements.txt
Flask==3.0.0
openpyxl==3.1.2
google-cloud-firestore==2.19.0
